refactor(core): single-home deepMerge in @bakin/core/merge
settings.ts's private copy and the adapter's exported copy were
semantically identical (same plain-object recursion condition, patch
wins, arrays/null replace wholesale) — verified before merging. The
adapter re-exports from core so its runtime-utils import path is
unchanged; the adapter already depends on @bakin/core.
